Switch skip connections from concatenation to addition
Matches original Cellpose U-Net design (concatenation=False, x=x+y).
Skip projection via 1x1 conv when encoder/decoder channels differ.
Also fix focal_loss to cast targets to long for cross_entropy.

+++ lkcellpose/utils/loss.py
@@ -75,7 +75,7 @@ class PanopticLoss(nn.Module):
         logits_valid = logits_flat[valid]
         targets_valid = targets_flat[valid]
 
-        ce = F.cross_entropy(logits_valid, targets_valid, reduction='none')
+        ce = F.cross_entropy(logits_valid, targets_valid.long(), reduction='none')
         pt = torch.exp(-ce)
 
         focal_weight = (1 - pt) ** self.focal_gamma
